requirements.txt
python-dotenv
openai
google-genai>=1.0.0
pandas
plotly
